docs(yard): fix debounce everywhere, feedback
