+++ documentation-site/pages/components/form-control.mdx
@@ -31,7 +31,7 @@ export default Layout;
 
 <Yard
   componentName="FormControl"
-  placeholderHeight={128}
+  placeholderHeight={136}
   {...formControlYardConfig}
 />
